requirements.txt
django==1.10
djangorestframework==3.6.2
pillow